requirements.txt
fastapi[all]==0.115.0
pydantic==2.9.2
uvicorn==0.31.0
jinja2==3.1.4
pydantic_settings==2.5.2
alembic==1.13.3
SQLAlchemy==2.0.35
python-jose==3.3.0
loguru==0.7.2
python-dotenv~=1.0.1
pytest~=8.3.3
aiohttp~=3.11.9
pytest-asyncio~=0.24.0
redis~=4.6.0